Slight Improvement on design
Made the text “THESE ARE ALL THE SKILLS YOU MAY NEED TO START YOUR
AMAZING JOURNEY IN THE WORLD OF OPEN SOURCE” color white, since the
grey shade was hard to read. The buttons in table were touching the
border which didn’t seem right, so added a bottom padding of 0.3em.